requirements.txt
pyahocorasick==2.1.0